Refactor: use lib scripts to return queue playlist
Playback should not be initiated from a subprocess; doing this means
that we have to duplicate this step in every queue builder script.

Instead, we refactor lets-watch to pass the playlist produced by the
queue builder, to the player. This means the script's only
responsibility is to print a playlist name (or an error message).