Major refactor; see commit comments
Yeah, sorry. I'd normally try to split these changes out more
thoughtfully, but this is a personal project and I got impatient.

Here's the rundown:

* Refactor scripts into plugins

Media provider scripts used to just run in a subshell. We now source
them and enforce a standard function interface. This enables additional
refactoring to enable greater flexibility, such as passing in
collections of searchable directories for folks with disjoint media
libraries.

All plugins must implement the function `lw_select_media` which prints a
newline-separated list of files.

* Centralize creation of playlist

The above refactors enable us to deduplicate temp playlist creation
logic, which was previously handled by delegate scripts individually.
This may seem redundant, but it enables work on additional planned
features, such as controlling player input by manipulating the generated
playlist. It also enables working with media players that don't handle
filenames with, e.g., spaces or other "irregularities."

* Overhaul config pattern

We no longer use environment variables for script configuration. This is
heavy handed and introduces application-specific baggage into
user-level scope. I'm not here to muck up your shell configs.

We now let the user set variables in a dedicated config script. The
interface is a bit cryptic, but hoping it meets the right balance in
terms of granularity and ease of use. Should be set-and-forget, so any
pains involved in setting up your config should only be a one-time cost.

* Simplify command line interface

We now call plugins by name; the default plugins have been renamed to
reflect the intended fluent syntax of the command (i.e., `lets-watch
local-media` to watch local media).

With this change, we can remove the command line switch for playlist
mode and adapt to the addition of custom plugins. We've added a switch
to enable setting a custom config file in case the user wants to
experiment efficiently with alternate configs.